Mutualize launchSecurityConfiguration method
